Move atomic_write to utils.py
