+++ fsspec/implementations/cache_metadata.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 
-import contextlib
 import os
 import pickle
-import tempfile
 import time
 from typing import TYPE_CHECKING
+
+from fsspec.utils import atomic_write
 
 if TYPE_CHECKING:
     from typing import Any, Dict, Iterator, Literal
@@ -204,24 +204,3 @@ class CacheMetadata:
     def update_file(self, path: str, detail: Detail) -> None:
         """Update metadata for specific file in memory, do not save"""
         self.cached_files[-1][path] = detail
-
-
-@contextlib.contextmanager
-def atomic_write(path: str, mode: str = "wb"):
-    """
-    A context manager that opens a temporary file next to `path` and, on exit,
-    replaces `path` with the temporary file, thereby updating `path`
-    atomically.
-    """
-    fd, fn = tempfile.mkstemp(
-        dir=os.path.dirname(path), prefix=os.path.basename(path) + "-"
-    )
-    try:
-        with open(fd, mode) as fp:
-            yield fp
-    except BaseException:
-        with contextlib.suppress(FileNotFoundError):
-            os.unlink(fn)
-        raise
-    else:
-        os.replace(fn, path)
